notebook/dns-hijack-detector/requirements.txt
requests==2.31.0
dnspython==2.4.2
networkx==3.2.1
matplotlib==3.8.4
